Fix UpdateDataObject tests
This commit is a first step towards updating the compliance
test suite to account for inconsistencies observed when
testing against the DOS server in ga4gh.dos.

In addition to fixing the request bodies of tests that call
the UpdateDataObject endpoint (which were erroneously missing
the 'data_object_id' parameter), this commit tweaks
test_full_data_object_update to only inspect attributes of
the data object changed in the test.